requirements.txt
selenium==4.29.*
webdriver-manager==4.0.*
SQLAlchemy>=2.0.38
pydantic==2.10.*
loguru
python-ulid